// src/editor/model/types.ts
export type TransactionSelection = EditorSelection | SelectionTarget;

export type EditorTransaction =
  | {
      kind: "keep-state";
    }
  | {
      kind: "replace-block";
      block: Block;
      blockId: string;
      listItemInsertedPath?: string;
      selection?: TransactionSelection | null;
    }
  | {
      kind: "replace-root";
      block: Block;
      rootIndex: number;
      selection?: TransactionSelection | null;
    }
  | {
      kind: "replace-root-range";
      count: number;
      replacements: Block[];
      rootIndex: number;
      selection?: TransactionSelection | null;
    }
  | {
      kind: "replace-selection-text";
      selection: EditorSelection;
      text: string;
    }
  | {
      kind: "set-selection";
      selection: EditorSelection;
    };